[#9120] Platform: Default alert route functionality
Summary:
Implemented default route (= default destination for notifications) behaviour.
Some details about it:
 - Default route is used when alerts are generated but the alert definition group (according to UI mocks - simply 'Alert configuration') doesn't have any destinations selected/configured;
 - Each route has flag 'Default route' which forces this route to be default; But each customer has only one default route, so if you mark another route as default, the previous default route will be unmarked (not used as default anymore).
 - For a new customer there is no default route until it specifies alert email recipients or creates a default alert route;
 - For existing customers we do a migration which creates a default route with alert email recipients and SMTP settings using existing customer's configuration settings (the default route will not be created if the customer still doesn't have recipients defined);

Added/fixed functionality not related to default routes:
 - It is impossible to create AlertRoute without receivers anymore;
 - It is impossible to delete receivers if it is the last receiver in any existing route;

TBD

Test Plan:
  TBD

Reviewers: amalyshev

Reviewed By: amalyshev

Subscribers: jenkins-bot

Differential Revision: https://phabricator.dev.yugabyte.com/D12145

+++ managed/src/main/java/com/yugabyte/yw/controllers/AlertController.java
@@ -46,8 +46,6 @@ public class AlertController extends AuthenticatedController {
 
   @Inject private AlertDefinitionGroupService alertDefinitionGroupService;
 
-  @Inject private AlertDefinitionService alertDefinitionService;
-
   @Inject private AlertService alertService;
 
   /** Lists alerts for given customer. */
@@ -290,8 +288,27 @@ public class AlertController extends AuthenticatedController {
   public Result deleteAlertReceiver(UUID customerUUID, UUID alertReceiverUUID) {
     Customer.getOrBadRequest(customerUUID);
     AlertReceiver receiver = AlertReceiver.getOrBadRequest(customerUUID, alertReceiverUUID);
-    log.info("Deleting alert receiver {} for customer {}", alertReceiverUUID, customerUUID);
 
+    List<String> blockingRoutes =
+        receiver
+            .getRoutesList()
+            .stream()
+            .filter(route -> route.getReceiversList().size() == 1)
+            .map(AlertRoute::getName)
+            .sorted()
+            .collect(Collectors.toList());
+    if (!blockingRoutes.isEmpty()) {
+      throw new YWServiceException(
+          BAD_REQUEST,
+          String.format(
+              "Unable to delete alert receiver: %s. %d alert routes have it as a last receiver."
+                  + " Examples: %s",
+              alertReceiverUUID,
+              blockingRoutes.size(),
+              blockingRoutes.stream().limit(5).collect(Collectors.toList())));
+    }
+
+    log.info("Deleting alert receiver {} for customer {}", alertReceiverUUID, customerUUID);
     if (!receiver.delete()) {
       throw new YWServiceException(
           INTERNAL_SERVER_ERROR, "Unable to delete alert receiver: " + alertReceiverUUID);
@@ -322,7 +339,13 @@ public class AlertController extends AuthenticatedController {
     AlertRouteFormData data = formFactory.getFormDataOrBadRequest(AlertRouteFormData.class).get();
     List<AlertReceiver> receivers = AlertReceiver.getOrBadRequest(customerUUID, data.receivers);
     try {
-      AlertRoute route = AlertRoute.create(customerUUID, data.name, receivers);
+      AlertRoute defaultRoute = AlertRoute.getDefaultRoute(customerUUID);
+      AlertRoute route = AlertRoute.create(customerUUID, data.name, receivers, data.defaultRoute);
+      if (data.defaultRoute && (defaultRoute != null)) {
+        defaultRoute.setDefaultRoute(false);
+        defaultRoute.save();
+        log.info("For customer {} switched default route to {}", customerUUID, route.getUuid());
+      }
       auditService().createAuditEntryWithReqBody(ctx());
       return YWResults.withData(route);
     } catch (Exception e) {
@@ -347,12 +370,27 @@ public class AlertController extends AuthenticatedController {
     Customer.getOrBadRequest(customerUUID);
     AlertRouteFormData data = formFactory.getFormDataOrBadRequest(AlertRouteFormData.class).get();
     AlertRoute route = AlertRoute.getOrBadRequest(customerUUID, alertRouteUUID);
+
+    if (route.isDefaultRoute() && !data.defaultRoute) {
+      throw new YWServiceException(
+          BAD_REQUEST,
+          "Can't set the alert route as non-default. Make another route as default at first.");
+    }
+
     List<AlertReceiver> receivers = AlertReceiver.getOrBadRequest(customerUUID, data.receivers);
     try {
+      AlertRoute defaultRoute = AlertRoute.getDefaultRoute(customerUUID);
       route.setName(data.name);
       route.setReceiversList(receivers);
+      route.setDefaultRoute(data.defaultRoute);
       route.save();
-
+      if (data.defaultRoute
+          && (defaultRoute != null)
+          && !defaultRoute.getUuid().equals(alertRouteUUID)) {
+        defaultRoute.setDefaultRoute(false);
+        defaultRoute.save();
+        log.info("For customer {} switched default route to {}", customerUUID, alertRouteUUID);
+      }
       auditService().createAuditEntryWithReqBody(ctx());
       return YWResults.withData(route);
     } catch (Exception e) {
@@ -364,6 +402,14 @@ public class AlertController extends AuthenticatedController {
   public Result deleteAlertRoute(UUID customerUUID, UUID alertRouteUUID) {
     Customer.getOrBadRequest(customerUUID);
     AlertRoute route = AlertRoute.getOrBadRequest(customerUUID, alertRouteUUID);
+    if (route.isDefaultRoute()) {
+      throw new YWServiceException(
+          BAD_REQUEST,
+          String.format(
+              "Unable to delete default alert route %s, make another route default at first.",
+              alertRouteUUID));
+    }
+
     log.info("Deleting alert route {} for customer {}", alertRouteUUID, customerUUID);
 
     AlertDefinitionGroupFilter groupFilter =
